Wireless_security_detector
In this project, I have designed a four-layer PCB consisting of two different boards. One of these cards takes on the task of a fake security camera, the other gives warnings thanks to the modules on it and provides various voltage outputs that you can use as an energy source.